backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
groq
tavily-python
pymupdf
python-docx
pillow
beautifulsoup4
lxml
python-multipart
httpx